requirements.txt
Flask
Pillow
playwright
pytz
